move to package + rename to *Holder + javadoc
