--- push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/cache/ServiceConstructor.java
@@ -1,5 +1,0 @@
-package org.jboss.aerogear.unifiedpush.message.cache;
-
-public interface ServiceConstructor<T> {
-    T construct();
-}
