--- push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/cache/ServiceDestroyer.java
@@ -1,5 +1,0 @@
-package org.jboss.aerogear.unifiedpush.message.cache;
-
-public interface ServiceDestroyer<T> {
-    void destroy(T instance);
-}
