+++ push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/serviceLease/ServiceDisposalScheduler.java
@@ -1,4 +1,4 @@
-package org.jboss.aerogear.unifiedpush.message.cache;
+package org.jboss.aerogear.unifiedpush.message.serviceLease;
 
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
@@ -9,6 +9,11 @@ import javax.annotation.PreDestroy;
 import javax.ejb.Singleton;
 import javax.ejb.Startup;
 
+/**
+ * Allows to scheduled instantiated services for disposal.
+ *
+ * Gracefully disposes services on a container shutdown.
+ */
 @Singleton
 @Startup
 public class ServiceDisposalScheduler {
@@ -16,11 +21,17 @@ public class ServiceDisposalScheduler {
     private ScheduledExecutorService scheduler;
     private long terminationTimeout = 10000L;
 
+    /**
+     * Creates a new scheduler on container startup
+     */
     @PostConstruct
     public void initialize() {
         scheduler = Executors.newSingleThreadScheduledExecutor();
     }
 
+    /**
+     * On container shutdown, immediately terminates all instantiated services that were scheduled for disposal.
+     */
     @PreDestroy
     public void terminate() {
         try {
@@ -31,13 +42,17 @@ public class ServiceDisposalScheduler {
         }
     }
 
+    /**
+     * Schedules a service instance for disposal if not used.
+     * @param reference a reference to service that will be disposed later
+     * @param delay a delay which need to pass before the reference can be disposed
+     */
     public void scheduleForDisposal(DisposableReference<?> reference, long delay) {
         synchronized (scheduler) {
             terminationTimeout = Math.max(delay + 2500L, terminationTimeout);
             scheduler.schedule(new DisposeTask(reference), delay, TimeUnit.MILLISECONDS);
         }
     }
-
 
     private static class DisposeTask implements Runnable {
