+++ push/sender/src/test/java/org/jboss/aerogear/unifiedpush/message/TestAbstractServiceCacheClustered.java
@@ -27,10 +27,10 @@ import javax.annotation.Resource;
 import javax.inject.Inject;
 import javax.jms.Queue;
 
-import org.jboss.aerogear.unifiedpush.message.cache.ApnsServiceCache;
-import org.jboss.aerogear.unifiedpush.message.cache.ServiceConstructor;
-import org.jboss.aerogear.unifiedpush.message.cache.ServiceDestroyer;
 import org.jboss.aerogear.unifiedpush.message.jms.util.JMSExecutor;
+import org.jboss.aerogear.unifiedpush.message.serviceLease.ApnsServiceHolder;
+import org.jboss.aerogear.unifiedpush.message.serviceLease.ServiceConstructor;
+import org.jboss.aerogear.unifiedpush.message.serviceLease.ServiceDestroyer;
 import org.jboss.aerogear.unifiedpush.test.archive.UnifiedPushArchive;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.OperateOnDeployment;
@@ -52,8 +52,8 @@ public class TestAbstractServiceCacheClustered {
     public static WebArchive archive() {
         return UnifiedPushArchive.forTestClass(TestAbstractServiceCacheClustered.class)
                 .withMessaging()
-                    .addPackage(org.jboss.aerogear.unifiedpush.message.cache.AbstractServiceCache.class.getPackage())
-                    .deleteClass(ApnsServiceCache.class)
+                    .addPackage(org.jboss.aerogear.unifiedpush.message.serviceLease.AbstractServiceHolder.class.getPackage())
+                    .deleteClass(ApnsServiceHolder.class)
                     .addClasses(MockServiceCacheForCluster.class)
                     .addAsWebInfResource("test-jms.xml")
                 .withMockito()
