+++ push/sender/src/test/java/org/jboss/aerogear/unifiedpush/message/sender/APNsPushNotificationSenderTest.java
@@ -29,8 +29,8 @@ import java.util.UUID;
 
 import org.jboss.aerogear.unifiedpush.api.iOSVariant;
 import org.jboss.aerogear.unifiedpush.message.UnifiedPushMessage;
-import org.jboss.aerogear.unifiedpush.message.cache.ApnsServiceCache;
-import org.jboss.aerogear.unifiedpush.message.cache.ServiceConstructor;
+import org.jboss.aerogear.unifiedpush.message.serviceLease.ApnsServiceHolder;
+import org.jboss.aerogear.unifiedpush.message.serviceLease.ServiceConstructor;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.mockito.invocation.InvocationOnMock;
@@ -73,8 +73,8 @@ public class APNsPushNotificationSenderTest {
         return baos.toByteArray();
     }
 
-    private ApnsServiceCache createMockApnsServiceCache() {
-        ApnsServiceCache apnsServiceCache = mock(ApnsServiceCache.class);
+    private ApnsServiceHolder createMockApnsServiceCache() {
+        ApnsServiceHolder apnsServiceCache = mock(ApnsServiceHolder.class);
         when(apnsServiceCache.dequeueOrCreateNewService(Mockito.anyString(), Mockito.anyString(), Mockito.any(ServiceConstructor.class))).thenAnswer(new Answer<ApnsService>() {
             @Override
             public ApnsService answer(InvocationOnMock invocation) throws Throwable {
